Migrate scenarios to DAML Scripts
In SDK 1.5.0 (and recent snapshots as the one used in this PR), DAML
Script is executed in DAML Studio and as part of `daml test` and we
are planning to deprecate scenarios so we are migrating over all
examples.

Happy to wait until SDK 1.5.0 before we merge this.

This is also a great example for why we did this: It allows us to kill
the whole duplication between the scenario and the script.